requirements.txt
numpy==1.24.3
torch==2.0.1
torchvision==0.15.2
matplotlib==3.7.1
scikit-image==0.21.0
pandas==2.0.2
tqdm==4.65.0
Pillow==9.5.0
scipy==1.10.1
tensorboard==2.13.0
torchinfo==1.8.0 